+++ modules/tinymce/src/themes/silver/test/ts/browser/editor/SilverEditorTest.ts
@@ -313,6 +313,9 @@ describe('browser.tinymce.themes.silver.editor.SilverEditorTest', () => {
                 ]
               }),
               s.element('div', {
+                classes: [ arr.has('tox-bottom-anchorbar') ]
+              }),
+              s.element('div', {
                 classes: [ arr.has('tox-statusbar') ]
               })
             ]